requirements.txt
# === BE Core — Smart Parking ===
# FastAPI + MQTT + DB

# Server
fastapi[standard]>=0.115.0
uvicorn>=0.30.0

# Database (Supabase PostgreSQL)
sqlalchemy>=2.0.48
psycopg2-binary>=2.9.11

# MQTT
paho-mqtt==1.6.1

# HTTP Client (gọi BE LPR)
httpx>=0.27.0

# AI Recognition (For Mock LPR and Testing) - STABLE VERSIONS
paddleocr==2.7.3
paddlepaddle==2.6.1
opencv-python>=4.6.0
numpy==1.26.4

# Utilities
python-dotenv>=1.0.0